// libs/features/feature360/src/lib/models/establishments/establishment-branches.ts
import { BilingualText } from '@gosi-ui/core';

export class EstablishmentBranches {
  registrationnumber: number = undefined;
  estnamearabic: string = undefined;
  estnameenglish: string = undefined;
  commercialregistrationnumber: number = undefined;
  locationarabic: string = undefined;
  locationenglish: string = undefined;
  licensenumber: string = undefined;
  status: BilingualText = new BilingualText();
}